refactored Recipe model

+++ backend/src/main/java/de/neuefische/backend/model/Temperature.java
@@ -5,12 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class Temperature {
 
-    private double number;
+    private BigDecimal number;
     private String unit;
 }
